Added role based permission and changed wrong function names in team.py

+++ routers/match_performance.py
@@ -1,9 +1,10 @@
 """Routes for match_performance"""
 
 from typing import Annotated
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter, Depends, HTTPException, Security
 from sqlmodel import Session
 from db.session import get_session
+from utils.authentication import get_current_user
 from crud import match_performance as crud
 from schemas import match_performance as schema
 
@@ -13,7 +14,9 @@ router = APIRouter(prefix="/match_performance", tags=["Match Performance"])
 
 @router.post("/create", response_model=schema.MatchPerformanceRead)
 async def create_match_performance(
-    new_match_performace: schema.MatchPerformanceCreate, session: Annotated[Session, Depends(get_session)]
+    new_match_performace: schema.MatchPerformanceCreate,
+    session: Annotated[Session, Depends(get_session)],
+    current_user: Annotated[dict, Security(get_current_user)],
 ):
     """Creates new match performance data"""
     return crud.create_match_perfomance(session, new_match_performace)
@@ -35,8 +38,14 @@ async def get_match_performance_by_id(id: int, session: Annotated[Session, Depen
 
 
 @router.delete("/delete", response_model=schema.MatchPerformanceRead)
-async def delete_match_perfomance_by_id(id: int, session: Annotated[Session, Depends(get_session)]):
+async def delete_match_perfomance_by_id(
+    id: int,
+    session: Annotated[Session, Depends(get_session)],
+    current_user: Annotated[dict, Security(get_current_user)],
+):
     """Deletes match performance by id"""
+    if current_user["role"] != "admin" and "superuser":
+        raise HTTPException(status_code=403, detail="Not enough privileges")
     delete_match_performanc_by_id = crud.delete_match_perfomance(session, id)
     if not delete_match_performanc_by_id:
         raise HTTPException(status_code=404, detail="Match peformance not found.")
